use command pattern
