+++ renku/cli/graph.py
@@ -18,27 +18,18 @@
 """PoC command for testing the new graph design."""
 
 import functools
-import shutil
 import sys
-from pathlib import Path
 
 import click
-from git import NULL_TREE, GitCommandError
 
 from renku.cli.utils.callback import ClickCallback
 from renku.cli.utils.click import CaseInsensitiveChoice
-from renku.core import errors
-from renku.core.commands.client import pass_local_client
 from renku.core.incubation.command import Command
-from renku.core.incubation.graph import GRAPH_METADATA_PATHS, export_graph, generate_graph
+from renku.core.incubation.graph import export_graph, generate_datasets_provenance, generate_graph
 from renku.core.incubation.graph import status as get_status
 from renku.core.incubation.graph import update as perform_update
-from renku.core.management.migrate import migrate
-from renku.core.models.provenance.datasets import DatasetProvenance
 from renku.core.models.workflow.dependency_graph import DependencyGraph
 from renku.core.utils.contexts import measure
-from renku.core.utils.migrate import read_project_version_from_yaml
-from renku.core.utils.scm import git_unicode_unescape
 
 
 @click.group(hidden=True)
@@ -175,124 +166,18 @@ _FORMATS = {
 
 @graph.command()
 @click.option("--format", type=CaseInsensitiveChoice(_FORMATS), default="json-ld", help="Choose an output format.")
-def export(format):
+@click.option("-d", "--dataset", is_flag=True, help="Include datasets.")
+def export(format, dataset):
     r"""Equivalent of `renku log --format json-ld`."""
     communicator = ClickCallback()
-    (export_graph().with_communicator(communicator).build().execute(format=_FORMATS[format]))
+    (export_graph().with_communicator(communicator).build().execute(format=_FORMATS[format], dataset=dataset))
 
 
 @graph.command()
 @click.option("-f", "--force", is_flag=True, help="Delete existing metadata and regenerate all.")
-@pass_local_client(requires_migration=True, commit=True, commit_only=GRAPH_METADATA_PATHS)
-def generate_dataset(client, force):
+def generate_dataset(force):
     """Create new graph metadata for datasets."""
-
-    commits = list(client.repo.iter_commits(paths=".renku/datasets/*"))
-    n_commits = len(commits)
-    commits = reversed(commits)
-
-    if force:
-        try:
-            client.datasets_provenance_path.unlink()
-        except FileNotFoundError:
-            pass
-    else:
-        if client.datasets_provenance_path.exists():
-            raise errors.OperationError("Dataset provenance file exists. Use `--force` to regenerate it.")
-
-    # Create empty dataset provenance file
-    client.datasets_provenance_path.write_text("[]")
-
-    datasets_provenance = DatasetProvenance.from_json(client.datasets_provenance_path)
-
-    for n, commit in enumerate(commits, start=1):
-        print(f"\rProcessing commits {n}/{n_commits}\r", end="", file=sys.stderr)
-
-        files_diff = list(commit.diff(commit.parents or NULL_TREE, paths=".renku/datasets/*"))
-        paths = [git_unicode_unescape(f.a_path) for f in files_diff]
-        deleted_paths = [git_unicode_unescape(f.a_path) for f in files_diff if f.change_type == "A"]
-
-        datasets, deleted_datasets = _fetch_datasets(client, commit.hexsha, paths=paths, deleted_paths=deleted_paths)
-
-        revision = commit.hexsha
-        date = commit.committed_datetime
-
-        for dataset in datasets:
-            datasets_provenance.update_dataset(dataset, client=client, revision=revision, date=date)
-        for dataset in deleted_datasets:
-            datasets_provenance.remove_dataset(dataset, revision=revision, date=date)
-
-    datasets_provenance.to_json()
+    communicator = ClickCallback()
+    (generate_datasets_provenance().with_communicator(communicator).build().execute(force=force))
 
     click.secho("\nOK", fg="green")
-
-
-def _fetch_datasets(client, revision, paths, deleted_paths):
-    from renku.core.models.datasets import Dataset
-
-    datasets_path = client.path / ".renku" / "tmp" / "datasets"
-    shutil.rmtree(datasets_path, ignore_errors=True)
-    datasets_path.mkdir(parents=True, exist_ok=True)
-
-    def read_project_version():
-        """Read project version at revision."""
-        try:
-            project_file_content = client.repo.git.show(f"{revision}:.renku/metadata.yml")
-        except GitCommandError:  # Project metadata file does not exist
-            return 1
-
-        metadata_path = datasets_path / "metadata.yml"
-        try:
-            metadata_path.write_text(project_file_content)
-            return int(read_project_version_from_yaml(metadata_path))
-        except ValueError:
-            return 1
-        finally:
-            metadata_path.unlink()
-
-    def copy_and_migrate_datasets():
-        existing = []
-        deleted = []
-
-        for path in paths:
-            rev = revision
-            if path in deleted_paths:
-                rev = client.find_previous_commit(path, revision=f"{revision}~")
-            identifier = Path(path).parent.name
-            new_path = datasets_path / identifier / "metadata.yml"
-            new_path.parent.mkdir(parents=True, exist_ok=True)
-            content = client.repo.git.show(f"{rev}:{path}")
-            new_path.write_text(content)
-            if path in deleted_paths:
-                deleted.append(new_path)
-            else:
-                existing.append(new_path)
-
-        try:
-            project_version = read_project_version()
-            client.set_temporary_datasets_path(datasets_path)
-            migrate(client, project_version=project_version, skip_template_update=True, skip_docker_update=True)
-        finally:
-            client.clear_temporary_datasets_path()
-
-        return existing, deleted
-
-    paths, deleted_paths = copy_and_migrate_datasets()
-
-    datasets = []
-    for path in paths:
-        dataset = Dataset.from_yaml(path, client=client)
-        # NOTE: Fixing dataset path after migration
-        original_identifier = Path(dataset.path).name
-        dataset.path = f".renku/datasets/{original_identifier}"
-        datasets.append(dataset)
-
-    deleted_datasets = []
-    for path in deleted_paths:
-        dataset = Dataset.from_yaml(path, client=client)
-        # NOTE: Fixing dataset path after migration
-        original_identifier = Path(dataset.path).name
-        dataset.path = f".renku/datasets/{original_identifier}"
-        deleted_datasets.append(dataset)
-
-    return datasets, deleted_datasets
