+++ renku/core/incubation/graph.py
@@ -17,6 +17,7 @@
 # limitations under the License.
 """Dependency and Provenance graph building."""
 
+import shutil
 from collections import defaultdict
 from pathlib import Path
 from typing import Dict
@@ -27,15 +28,18 @@ from renku.cli.update import execute_workflow
 from renku.core import errors
 from renku.core.incubation.command import Command
 from renku.core.management.config import RENKU_HOME
+from renku.core.management.migrate import migrate
 from renku.core.management.repository import RepositoryApiMixin
 from renku.core.models.entities import Entity
 from renku.core.models.provenance.activities import Activity as ActivityRun
 from renku.core.models.provenance.activity import ActivityCollection
+from renku.core.models.provenance.datasets import DatasetProvenance
 from renku.core.models.provenance.provenance_graph import ProvenanceGraph
 from renku.core.models.workflow.dependency_graph import DependencyGraph
 from renku.core.models.workflow.run import Run
 from renku.core.utils import communication
 from renku.core.utils.contexts import measure
+from renku.core.utils.migrate import read_project_version_from_yaml
 from renku.core.utils.scm import git_unicode_unescape
 
 GRAPH_METADATA_PATHS = [
@@ -68,7 +72,7 @@ def _generate_graph(client, force):
             pass
     else:
         if client.has_graph_files():
-            raise errors.OperationError("Graph files exist. Use --force to regenerate the graph.")
+            raise errors.OperationError("Graph files exist. Use ``--force`` to regenerate the graph.")
 
     create_empty_graph_files()
 
@@ -191,9 +195,17 @@ def update():
     return command.require_migration().with_commit(commit_if_empty=False).require_clean()
 
 
-def _export_graph(client, format):
+def _export_graph(client, format, dataset):
     """Output graph in specific format."""
+    if not client.provenance_graph_path.exists():
+        raise errors.ParameterError("Graph is not generated.")
+
     pg = ProvenanceGraph.from_json(client.provenance_graph_path, lazy=True)
+
+    if dataset:
+        if not client.datasets_provenance_path.exists():
+            raise errors.ParameterError("Dataset provenance is not generated.")
+        pg.rdf_graph.parse(location=str(client.datasets_provenance_path), format="json-ld")
 
     return format(pg.rdf_graph)
 
@@ -218,3 +230,120 @@ def _get_modified_paths(client, plans_usages):
                 modified.add(plan_usage)
 
     return modified, deleted
+
+
+def _generate_datasets_provenance(client, force):
+    """Generate datasets provenance metadata."""
+    commits = list(client.repo.iter_commits(paths=".renku/datasets/*"))
+    n_commits = len(commits)
+    commits = reversed(commits)
+
+    if force:
+        try:
+            client.datasets_provenance_path.unlink()
+        except FileNotFoundError:
+            pass
+    else:
+        if client.datasets_provenance_path.exists():
+            raise errors.OperationError("Dataset provenance file exists. Use ``--force`` to regenerate it.")
+
+    # Create empty dataset provenance file
+    client.datasets_provenance_path.write_text("[]")
+
+    datasets_provenance = DatasetProvenance.from_json(client.datasets_provenance_path)
+
+    for n, commit in enumerate(commits, start=1):
+        communication.echo(f"\rProcessing commits {n}/{n_commits}\r")
+
+        files_diff = list(commit.diff(commit.parents or NULL_TREE, paths=".renku/datasets/*"))
+        paths = [git_unicode_unescape(f.a_path) for f in files_diff]
+        deleted_paths = [git_unicode_unescape(f.a_path) for f in files_diff if f.change_type == "A"]
+
+        datasets, deleted_datasets = _fetch_datasets(client, commit.hexsha, paths=paths, deleted_paths=deleted_paths)
+
+        revision = commit.hexsha
+        date = commit.committed_datetime
+
+        for dataset in datasets:
+            datasets_provenance.update_dataset(dataset, client=client, revision=revision, date=date)
+        for dataset in deleted_datasets:
+            datasets_provenance.remove_dataset(dataset, revision=revision, date=date)
+
+    datasets_provenance.to_json()
+
+
+def _fetch_datasets(client, revision, paths, deleted_paths):
+    from renku.core.models.datasets import Dataset
+
+    datasets_path = client.path / ".renku" / "tmp" / "datasets"
+    shutil.rmtree(datasets_path, ignore_errors=True)
+    datasets_path.mkdir(parents=True, exist_ok=True)
+
+    def read_project_version():
+        """Read project version at revision."""
+        try:
+            project_file_content = client.repo.git.show(f"{revision}:.renku/metadata.yml")
+        except GitCommandError:  # Project metadata file does not exist
+            return 1
+
+        metadata_path = datasets_path / "metadata.yml"
+        try:
+            metadata_path.write_text(project_file_content)
+            return int(read_project_version_from_yaml(metadata_path))
+        except ValueError:
+            return 1
+        finally:
+            metadata_path.unlink()
+
+    def copy_and_migrate_datasets():
+        existing = []
+        deleted = []
+
+        for path in paths:
+            rev = revision
+            if path in deleted_paths:
+                rev = client.find_previous_commit(path, revision=f"{revision}~")
+            identifier = Path(path).parent.name
+            new_path = datasets_path / identifier / "metadata.yml"
+            new_path.parent.mkdir(parents=True, exist_ok=True)
+            content = client.repo.git.show(f"{rev}:{path}")
+            new_path.write_text(content)
+            if path in deleted_paths:
+                deleted.append(new_path)
+            else:
+                existing.append(new_path)
+
+        try:
+            project_version = read_project_version()
+            client.set_temporary_datasets_path(datasets_path)
+            migrate(client, project_version=project_version, skip_template_update=True, skip_docker_update=True)
+        finally:
+            client.clear_temporary_datasets_path()
+
+        return existing, deleted
+
+    paths, deleted_paths = copy_and_migrate_datasets()
+
+    datasets = []
+    for path in paths:
+        dataset = Dataset.from_yaml(path, client=client)
+        # NOTE: Fixing dataset path after migration
+        original_identifier = Path(dataset.path).name
+        dataset.path = f".renku/datasets/{original_identifier}"
+        datasets.append(dataset)
+
+    deleted_datasets = []
+    for path in deleted_paths:
+        dataset = Dataset.from_yaml(path, client=client)
+        # NOTE: Fixing dataset path after migration
+        original_identifier = Path(dataset.path).name
+        dataset.path = f".renku/datasets/{original_identifier}"
+        deleted_datasets.append(dataset)
+
+    return datasets, deleted_datasets
+
+
+def generate_datasets_provenance():
+    """Return a command for generating dataset provenance."""
+    command = Command().command(_generate_datasets_provenance).lock_project()
+    return command.require_migration().with_commit(commit_only=GRAPH_METADATA_PATHS)
