+++ tests/core/commands/test_client.py
@@ -60,6 +60,7 @@ def test_safe_class_attributes(tmpdir):
         "CACHE",
         "CONFIG_NAME",
         "DATASETS",
+        "DATASET_PROVENANCE",
         "DATA_DIR_CONFIG_KEY",
         "DEPENDENCY_GRAPH",
         "DOCKERFILE",
@@ -82,6 +83,7 @@ def test_safe_class_attributes(tmpdir):
         "_CMD_STORAGE_UNTRACK",
         "_LFS_HEADER",
         "_global_config_dir",
+        "_temporary_datasets_path",
     ]
 
     client1 = LocalClient(str(tmpdir.mkdir("project1")))
